Fixes #33. Fixes #32. Fixes other bugs (see below)
There was a bug where a player "ghost" was left after it was stepped on by an actual ghost. This is because the game code treated Pac-man as an object to be saved or "picked up", but then all characters were teleported by reset_positions(). This ended up leaving the corpse of pac-man either where pac-man died or in the ghost house after the teleported ghosts takes its first step.