Constants now use ConstMapping like Space Invaders
I updated the code to use ConstMapping like Space Invaders does.

I did this in the process of trying to solve Issue #37, which is that when I touch map_array[][] in an LP program it crashes with the errors quoted in #37. As far as I can tell, I am creating and providing the map to users in bot_vars exactly the same way I do with Space Invaders.