Track ghosts when edible (Fixes #38)
This commit fixes Issue #38.

I fixed it by creating new sprites for each ghost for when they are edible. The Ghost object has a normal `char` and an `edible_char` for when it is edible. The appropriate sprite is mapped depending on its personal vulnerability level (once eaten, ghosts are no longer vulnerable even if others are). The sensors (e.g.) inky_x and inky_y search for either INKY or EDIBLE_INKY and return whichever is visible. The point sensors (e.g.) sense_n will return GHOST regardless of which ghost or its edibility.