Fixes #41
This bug must have existed for a long time.

Now there is a counter for fruit's visibility -- fruit always lasts for 50 turns. When the counter hits zero, it is erased. The counter is set to zero when Pac-Bot eats the fruit.

When initialized (new game or new level), self.fruit_visible is set to -1 so that it is not visible and not to be erased.